feat(TT-19): moved to constant file
