+++ src/components/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Navigate } from "react-router-dom";
 import { useAuth } from "@/lib/auth";
-import { ROUTES } from "@/routes";
+import { ROUTES } from "@/lib/constants";
 
 type Props = {
   children: React.ReactNode;
