+++ src/lib/constants.ts
@@ -1,0 +1,4 @@
+export const ROUTES = {
+  HOME: "/",
+  LOGIN: "/login",
+};
